Add basic `rcfiles` manager and initial installer
The `rcfiles` binary is for managing this repository from a place
external to it, as it is located along the $PATH specified by both
~/.bashrc and ~/.zshrc. It is executed via `rcfiles <subcommand>`. The
following subcommands are currently available:

 * relink - Move existing dot-rc files from $HOME (or wherever they are
   linked to) into the current directory, and relink them there. This
   command can be used to import existing differences from current rc
   files, or discarding them, using standard git commands.
 * autoconfig - Set up the repository-local git configuration with
   necessary settings for proper functionality. Currently this means
   applying a filter so that certain semi-private information (name and
   email address) are not committed to the repository by accident.
 * status - Lets you know whether there are any differences between the
   working copy and the repository. It is a shorthand for both running
   `git status` in the rcfiles repository as well as its output.
 * update - Attempts to pull changes from the remote rcfiles repository.

`rcfiles` is also written in such a way that it can be included as a
Perl module, though it does not return a true value and this must be
accounted for. All these subcommands are methods in the RCFiles package.

The installer script is written in a way that it can both be locally
executed, as well as piped into `perl` and executed from a download (ie,
that way tons of people use but almost everyone else hates). Everybody
wins! Soon I'll write a README which details its usage.