Make change-directory behavior more consistent
Since there was a lot of code-copying going on, there's now a sub to
handle changing the directory and behave consistently across the whole
script. While nothing yet uses it, it takes an optional argument, which
should be a true value of changing the directory is not necessarily a
fatal error.