Add Syntastic/perl-mauke, remove csv, move php.vim
Syntastic should help detect problems before exiting vim and trying to
run scripts. Hooray! For now, due to a couple problems with installed
library locations, I've made it so that PHP will not be actively
checked; I should fix this and update the file eventually.

The CSV plugin wasn't behaving properly, and so it's gone. It's possible
the wrong CSV plugin was loaded, but I work with CSV rarely enough that
it wasn't currently worth the trouble to fix it. I'd already commented
it out, but now it's just gone.

I've also re-enabled the perl-mauke updated syntax bundle. I had it for
a while, but had disabled it while I was trying to figure out whether it
was the cause of issues with lack of HEREDOC indenting. It was not, and
unfortunately I haven't figured out what the problem really is. However,
in the meantime, best not to throw the baby out with the bathwater.

I also fixed indentation in one random line. That was very important.