Fix links in README (again), and mention `reshell`
So, in commit a703963828ca26029fe50de0a58e949ce8f29461 I thought I'd
gotten a handle on this whole link thing. It turns out, though, that I'm
not very good at Markdown, after all... (._. )

(That was a disappointed/downcast face.)

I've also mentioned the new `reshell` script in the README, as it's
another way you can interact with the system, so the old paragraph was
a big ugly pile of lies, terrible lies.