Properly document what autoconfig does in install
While this doesn't impact any functionality, I don't like having
outdated documentation (even if that's just a comment). If this gets
much more complicated, I'll probably just take out the existing
documentation comment and say "go look at the file if you're curious",
because it becomes increasingly likely I'll forget to keep updating it,
and it'd be tedious to do so.