Make subs more consistent/shorter in `rcfiles`
New subs:
  * __good - Print a positive/informational message. Return __success__.
  * __bad - Print a negative/nonfatal error. Return nothing.
  * __die - Print a fatal error. Return __failure__.

Renamed subs:
  * __print__   -> __print
  * __warn__    -> __warn
  * __chdir__   -> __chdir
  * __findsub__ -> __findsub

The general rationale is that private subs should be prefixed with `__`,
but the suffix of `__` is unnecessary, and could instead be used to
indicate a constant. `__main__` is an exception to this as it is a
special method that should only ever be called once.

Additionally, the way error messages and informational messages are
printed could change at any time, and it would be much better to have it
be done through a consistent method so that, should I choose to do so,
it will only need to happen in once place. Also, being able to print a
fatal error and return on the same line is a nice new benefit.