Add `reset` subcommand to `rcfiles`
`reset` will clear all of the modifications to the local copy of the
rcfiles repository. This includes clearing out the staging index from
git, removing untracked (but not ignored) files, and resetting all
changes to the latest HEAD version. There is also an informative comment
on the `rcfiles status` output to showcase this option.