Update some Vim PostgreSQL interactions
Sometimes I write `DO` statements into my SQL files as part of
migrations or simply when using "vim-slime" to edit large and
complicated queries in another tmux pane, and so it's helpful to have
the code highlighted properly for PLPERL in this case as well, and not
just function definitions.

The other update changes the mapping for sending a PostgreSQL query in
the background to prefix it with the <Leader> key. This command wasn't
in my memory very deeply and it's very uncommon that I use it, so I
figure it doesn't need to be overwriting any built-in commands like
that, when this sort of thing is precisely what <Leader> is useful for.