Ensure updated `rcfiles` configuration is reloaded
The code that runs after `rcfiles update` would previously only
guarantee that anything that somehow became desynchronized from the
pre-update version would actually become properly relinked or
reconfigured according to said pre-update version, because it would not
have any updates to the `rcfiles` code yet. However, the point of the
code that does that was intended to add links and configurations from
the post-update version of the code.

Therefore, now there is an __exec() sub which will take a single
argument: the name of a subcommand to run against the current version of
the files. It shifts the resulting code right by 8 bits per the perldoc
and returns that, so that ideally it will be almost indistinguishable
from the way it would worked if you called the subroutines directly,
except for the increased memory and CPU footprint.

The two command in question, `relink` and `autoconfigure`, have been
switched over to this method and they appear to be working properly.
Now, `rcfiles update` will do all of the reloading that it's possible to
do without user intervention (shell aliases/etc. still need to be
manually reloaded with the `reshell` command function).