Add inline help for all commands
Now, when a command is left off, a usage text is printed, listing all
commands, with a summary. If you run `rcfiles help`, it will do the
same, but exit with a success status instead of failure.

If you run `rcfiles help $COMMAND`, then a more full-fledged help for
$COMMAND will be printed.

All of this help is simply created inline proper-designed comments - any
comment occurring immediately before a subroutine definition will create
help text. The first line will be the summary, a single line will be
skipped, and every line after that will be the full help text.