Make `rcfiles relink` report success/failure
In addition to properly reporting whether or not it was able to do
everything it is intended to do, the `relink` subcommand is now also
more explicit in what it's doing at any given time, or if it doesn't do
anything at all. This is much more like the other `rcfiles` subcommands.