Populate the location list from Syntastic by def.
I'll probably add some shortcuts soon, but in the meantime, this allows
me to navigate to errors caught by Syntastic with the :lprev and :lnext
commands, which is handy. The reason it isn't on by default is that it
could conflict with other plugins that want to use the location list,
but I don't think I've got any of those at the moment, so we should be
pretty safe here.