Silence certain Perl::Critic warnings on rcfiles
I finally enabled Perl::Critic, only to find that subroutine prototypes
are discouraged! That, in turn, is discouraging! I can understand the
long-term reasoning, but for now this script is a one-off binary and I
think the layout works better, so we're ignoring that.

Additionally, the `no strict 'refs';` line in `bin/rcfiles` is necessary
for what I'm doing there, and at the very least very intentional and
very limited, so I've silenced the one warning generated by that line.