Modify `rcfiles gitconfig` to remove old keys
Since the local version of `alias.full-diff` is gone in favor of the
global version, we should remove the local one from rcfiles (where it's
almost definitely going to be used), as the order of config file
precedence would mean that the old alias would supersede the new one.

Now all stale keys can be auto-removed from any given update.