Handle earlier versions of Vim better
In older versions of Vim, "matchadd" isn't available and so we need to
have a special case for turning off anything that uses it, or else we'll
get a bunch of errors on startup every singly time.

This is largely for working on older RedHat/CentOS boxes, or other
systems which have significantly older versions of Vim ("matchadd" was
introduced in 7.2).