Add `git force-merge` alias and script
I was doing this by hand for a while; it's probably actually bad
practice, but essentially I would find that I'd written a bunch of
changes, and in the midst of separating them all out into discrete
commits, I'd forget some modifications that really *should* have been
grouped together (something that could conceivably break the process for
its absence), but I'd have created many commits in the meantime. So, I'd
force a merge to an earlier commit on the same branch, which looks a
little bit weird but is actually quite allowed because of how git works.
Weird, huh?

Oftentimes, this would happen before I pushed the changes, so I really
*should* have been using `git rebase` to go back and rewrite history,
but I'm strangely enough more comfortable with this commands. Though it
would probably be *better* to `rebase`, I find this works better in the
broad sense of "maybe I pushed the commits up, maybe I didn't", because
in the former case, I shouldn't be rewriting the history for my trivial
mistakes.

In theory, there could be multiple parent commits, and I do think I've
even done that once, but at this point the command only allows for one
commit "tree-ish" object to be specified. I say SHA, but again (because
of how git works), it could conceivably be another branch. However, if
it's another branch or other reference, forcing the merge seems
unnecessary, as you could simply do `git merge --no-ff <BRANCH>`. But,
who am I to argue?