Fix a (sorta) bug in `git full-diff`
In theory, I intended it such that `git full-diff` could be used like
`git diff` where you could specify a file and just diff that. I so
rarely used it this way, that I didn't realize this was broken.
Technically, it still kind of is because it doesn't actually diff the
file if it's specified as an argument but not in the index. Again, I
rarely ever use this, so I'm not in a huge rush to fix that.