Switch out perlbrew/RVM for (pl|rb)env
The config actually supports both now, but the assumption is that we'll
prefer plenv/rbenv to perlbrew/RVM, and a warning will appear on every
login where it appears (to some surface level) like there's both
perlbrew+plenv or RVM+rbenv on the system.

As far as I'm able to tell, the only way things will *really* blow up is
if we try to configure to use both; the admonitions to make sure they
don't exist on your system simultaneously seem overblown.

That being said, there's no *reason* to have both at the same time, and
better safe than sorry.