Replace `function!` with `function` in init.vim
In general, we might want to use `function!` because we'd rather just
ensure that the function is defined, and don't care if the file was
already sourced. However, there are specific guards in place to prevent
these files from reloading, and these will only increase in number, so
if the functions are getting redefined, that means something else is
going wrong, so we'll probably want to be aware of it.