Replace project/dir-specific .vimrc loading code
This code is a little bit more elegant, but the general idea remains the
same: directory-specific .vimrc files get loaded, as do ones specific to
git projects (located either in the root directory of the project in
.vimrc, or in .git/vimrc).

The order of loading has changed from before, however: the
directory-specific configs are loaded *before* the project-specific
ones. This is because I may eventually disable further rc file loading
with a function call any given config can issue, and it seems like one
for a directory should be able to override one for a project and halt
the project-specific ones.