Move general settings/filetype mods to vim/ folder
The FileType autocommand handling is a little bit more elegant now, in
that you can simply define a function for a given FileType in the
vim/filetypes.vim file, named after that FileType, and its behavior will
be automatically loaded. This allows multiple commands to be run off one
FileType action.

The settings are a bit more granularized in terms of how they're broken
apart; hopefully this makes it easier to understand, and not more
complicated. As far as I can recall, no specific setting was actually
changed at this point, just broken apart.