Rewrite Vundle compatibility layer
It turns out that I had a completely wrongheaded conception of how the
upgrade path to newer versions of Vundle would work: I thought there was
a deprecated repo named 'vundle', and an upgraded repo called
'Vundle.vim', and ne'er the twain should meet. However, the 'Vundle.vim'
repo is simply the newer version, and 'vundle' is an *alias* to that, so
that you can have an upgraded version that's still in the old path.

So, I've rewritten the compatibility layer, because there are three
different things to account for:

    1. Is the path up-to-date?
    2. Is there a #begin()/#end() pair of commands?
       - This occurred historically after, and definitely separate from,
         the move to 'Plugin' as the preferred command name family for
         interacting with Vundle.
    3. Are there 'Plugin' commands at all?

There are now independent mechanisms in place to handle a "no" answer to
any of these three questions, none of which depend on each other, and
each of which leaves some marker information in g:myVim.vundle.compat,
which before was a simple Number but is now a Dictionary.