Change git push.default from 'simple' to 'nothing'
Though 'simple' is the default in newer versions of git, and
intermediate versions of git advise you to do this, it turns out that I
use systems where the version of git is old enough that this causes
problems for me because 'simple' as an option doesn't exist in those
versions at all.

Since I'm in the habit anyway of specifying refspecs to where I want to
push/pull to/from as I frequently work with multiple remotes and even
occasionally keep local branch names different than those on the remote
server (because I'm a equal parts clownshoes and a jerk), I don't really
mind being "forced" to.