Alter strip-ini-section to use STDIN instead of fn
Speaking of old git versions, the '%f' expansion in the filter.x.clean
option is newer than some of the versions of git I use, and in this
particular case, is actually unneeded - the script already gets fed the
contents into STDIN, so all we really need to do is just read that. It
makes the 'strip-ini-section.pl' file a bit less reusable, but then, I'm
not really using it anywhere else, so whatever I guess.