Add `lsa` alias, make `ls` not list all files
I recently have found myself more and more writing `\ls` with a custom
set of flags, just to strip out the dot files that show up in the
listing, so I've gone ahead and made it two separate command aliases.
This way, it's fairly easy to add the hidden files back in, but they
aren't there by default.