Update colors to use a light theme now
This all came about as a result of wanting to try out the PaperColor
theme as mentioned in UseVim: http://usevim.com/2015/05/22/paper-color/

I wound up deciding that it looked decent enough, but I'm not a huge fan
of switching the entire color of my terminal because I opened one
particular program, so I wound up updating the color scheme for
everything else as well: iTerm, zsh, ls, you name it.

Because the "cloud" theme doesn't really work on light backgrounds (in
fact, almost none of them do, it seems light terminals are not the most
beloved), I picked a pretty decent replacement ("sammy") and slightly
tweaked it. This involved setting a custom $ZSH_CUSTOM directory, so it
could be managed here.

The tmux configuration was slightly less cumbersome, but now I have to
use color *numbers* (and the non-American spelling of "colour" as wel)
to specify the colors instead of names.

And, the most complicated bit is probably the $LSCOLORS/$LS_COLORS. I
would not have been able to figure any of that out without the online
generator by @ggreer: http://geoff.greer.fm/lscolors/o

I'm still not sold on highlighting for any *common* thing like a
directory, but we'll see how it goes. I may just wind up trashing the
light theme wholesale.