Add `short-dirname` script
This can be loaded either as a library or run as a script, but the long
and short is that it shortens long directory names in a way that could
be referenced by a shell script. It can be used by tmux or other things
that need to display path names to the user so that they don't take up
too much screen real estate.