Add scripts to automatically update the system
With a couple of simple configuration lines, it's possible to now have
the system automatically update various system components.

If something is specific to a site, it can be put into
$RCDIR/site/autoupdate-config, or ~/.autoupdate-config. Items are
defined as lines with "item.key=value". The only keys currently known
are "command" and "expiration". The command is a line of bash code that
will be eval'd, so it shouldn't be written as though it's safe. The
"expiration" is how many seconds should pass between executions of that
command.

The "expiration" value is not required, if it isn't given, then the
default value is 2 weeks.

The update scripts that are already defined are an "autoupdate" one
which runs every 30 minutes and which if expired will run all other
update scripts, and an "rcfiles" on which will run once a week.