Ensure "reshell" function exists before calling it
In the autoupdate script, it's running underneath a different shell
process, and that process doesn't have the "reshell" function loaded, so
we need to make sure to manually load it before trying to run it.

I tried simply moving it so that "reshell" is its own executable script,
but that makes it no longer work from the current shell, as it will be
executing in a subprocess (basically, the same problem that requires
this modification in the first place).