Add command to auto-update vim plugins
This command presumes the presence of 'tmux' to run the plugin update in
the background, as it (naturally) would otherwise take over the entire
terminal for some amount of time, and that just won't do.

There are a few unfortunate side-effects of this:

  * If tmux isn't present, the command will always fail, and thus every
    time a new terminal is opened after 30 minutes, it will continue to
    try and fail to update the plugins.
  * If tmux is present, the command will always succeed, regardless of
    whether or not the command was successful in updating the plugins.
    An example of a potential reason it could fail is if there's no
    network.
  * If the command hangs waiting for user input (as naturally vim
    doesn't know whether or not a user is actually meant to watch it
    running), there's no way to know aside from the fact that the
    'vim-update' tmux session will last longer than it should.

At some point I may write a more elaborate script which is capable of
handling all these variables, but for now, this should at least mostly
do the trick.