Add 'CtrlP' and 'Tabular' plugins for vim
One is just something I hear people talking about *all* the time, so I
figured why the heck not.

The other was suggested by Tim Pope (or at least, he said it's one of
the ones he likes that isn't one he himself wrote, and that's good
enough for me).