Change <Leader> to <Space> in Vim
I'm certainly not using these mappings enough, and ";" may be partly to
blame. I'm not a huge fan of that as a leader. Any time I've done
site-local mappings, it's usually been from the spacebar. I think it's a
bit more of a natural fit. We'll see if I actually use any of these
mappings now.