Handle tmux screwing up utf8
Technically, I think, based on the commit messages and various issues
that were opened in the tmux blog, this is technically Apple's fault?
Or, alternately, Powerline's.

But, it worked fine, and then an upgrade to tmux broke something, so I'm
blaming tmux, cause it was totally in the realm of something that wasn't
broke that they decided to fix.

In short: they had a table of unicode characters which worked, but
yanked it in this commit:
https://github.com/tmux/tmux/commit/26945d7956bf1f160fba72677082e1a9c6968e0c

At this point, I only know it's broken inside of tmux 2.2, but if I find
it broken elsewhere, I guess I'll patch this later.