Add "git is-merged" command
This command simply wraps "git merge-base --is-ancestor" so that I don't
have to remember that command, which is not named very well, if I do say
so myself.

Essentially, I wanted a quick way to see if I'd already merged in a
particular commit when working between several remotes and branches at
the same time.