Add hjkl pane navigation to tmux configuration
I...may have broken one of my arrow keys. So, to make it easier to
navigate panes again in tmux, I've added standard "hjkl" (LDUR)
navigation to the configuration. It overrides a built-in binding
(last-window on "l"), but the rest weren't on any command, and I never
use the "last-window" command anyway.