Allow vim autoupdate to start from tmux sessions
Because the Vim autoupdate works by spawning a new tmux session, it
needs to have the $TMUX variable be unset in case it's being spawned as
a result of a new session starting inside of a tmux panel. Since the
tmux session is starting detached, it isn't really nested (not that
there's anything wrong with that), but tmux requires the extra step in
order to make sure it's what we want. Which, it is.