Add 'root' and 'rmux' session wrappers
The first merely requests to immediately jump to the `root` account by
way of `sudo` on the remote system. "Don't do things as root!" I hear
the people exclaim. I know, I know. Sometimes you just have to take the
safety plate off to get at those tantalizing dangerously-whirling
blades.

The other command is a specific case of using the `root` command to
request to create-or-attach to a specific remote session in that `root`
account. This is more likely propping the door to the nuclear reactor
chamber so we can get in quickly two make sure the drinking bird is
still tapping the "Yes" button to prevent a total meltdown.