Update pg-switch directories
I never really cared for the several-postgres-installations-inside-my-
home-directory method of doing this, so now I've made it so that
everything's inside of a single `.pgenv` directory, which makes it a
little bit easier to manage. There's not really a `pgenv` setup akin to
`rbenv` or `plenv` or any others, but it was as good a name as any.