Add more "wip" git commands
These theoretically don't do anything that wasn't already being done,
but the fact that they're all broken up now means that we can easily run
any given portion of the commands on their own. This is particularly
useful if, for instance, we get partially through the process, and
something fails, and so we want to start from that point and move on,
instead of needing to start from scratch, or remember the appropriate
commands by hand.